- Fix issue with comments on Message Detail screen sometimes not loading when opened via push notification (would get stuck showing Loading message) - Convert MessageEventModel and MessageRecipientModel to use callbacks instead of delegate - Fix crash issue that occurred on debug build when attempting to call TeleMed without a phone number being set

+++ iOS/AppDelegate.m
@@ -114,7 +114,8 @@
 			UNUserNotificationCenter *userNotificationCenter = [UNUserNotificationCenter currentNotificationCenter];
 			
 			[userNotificationCenter setDelegate:self];
-			[userNotificationCenter requestAuthorizationWithOptions:(UNAuthorizationOptionSound | UNAuthorizationOptionAlert | UNAuthorizationOptionBadge | UNAuthorizationOptionCarPlay) completionHandler:^(BOOL granted, NSError * _Nullable error) {
+			[userNotificationCenter requestAuthorizationWithOptions:(UNAuthorizationOptionSound | UNAuthorizationOptionAlert | UNAuthorizationOptionBadge | UNAuthorizationOptionCarPlay) completionHandler:^(BOOL granted, NSError * _Nullable error)
+			{
 				if (granted)
 				{
 					dispatch_async(dispatch_get_main_queue(), ^
@@ -259,7 +260,7 @@
 	if (timeoutEnabled && ! [self isCallConnected])
 	{
 		// Delay logout to ensure application is fully loaded
-		dispatch_after(dispatch_time(DISPATCH_TIME_NOW, 0.5 * NSEC_PER_SEC), dispatch_get_main_queue(), ^
+		dispatch_after(dispatch_time(DISPATCH_TIME_NOW, (int64_t)(0.5 * NSEC_PER_SEC)), dispatch_get_main_queue(), ^
 		{
 			[authenticationModel doLogout];
 		});
