+++ iOS/Shared/Controllers/LoginSSOViewController.m
@@ -118,7 +118,7 @@
 		[self updateWebViewLoading:YES];
 		
 		// Delay is here is required because there is a slight delay between device going back online and requests actually going through
-		dispatch_after(dispatch_time(DISPATCH_TIME_NOW, 2.0 * NSEC_PER_SEC), dispatch_get_main_queue(), ^
+		dispatch_after(dispatch_time(DISPATCH_TIME_NOW, (int64_t)(2.0 * NSEC_PER_SEC)), dispatch_get_main_queue(), ^
 		{
 			[self initLogin];
 		});
