+++ iOS/Shared/Models/Core/Model.m
@@ -83,7 +83,7 @@
 			if (callback != nil)
 			{
 				// Delay callback to ensure that activity indicator has finished dismissing
-				dispatch_after(dispatch_time(DISPATCH_TIME_NOW, 0.25 * NSEC_PER_SEC), dispatch_get_main_queue(), ^
+				dispatch_after(dispatch_time(DISPATCH_TIME_NOW, (int64_t)(0.25 * NSEC_PER_SEC)), dispatch_get_main_queue(), ^
 				{
 					callback();
 				});
